custom hooks control de like
