+++ src/components/PhotoCard/index.js
@@ -1,16 +1,17 @@
 import React, { useEffect, useRef, useState, Fragment } from 'react'
 import { Button, ImgWrapper, Img, Article } from './style'
 import { MdFavorite, MdFavoriteBorder } from 'react-icons/md'
+import { UsearScreen } from '../../hooks/useLocalStorage'
 
 const IMAGE_DEFAULT = "https://i.postimg.cc/y8qgfJ0z/pexels-alexandra-holbea-6056878.jpg"
 
 function useLocalStorage(key, initialValue) {
     //custom hooks
-    const [value, setValue] = useState(() => {
-        //obtener el like del localstorage
+    //obtener el like del localstorage
+    const [values, setValue] = useState(() => {
         try {
-            const like = window.localStorage.getItem(key)
-            return value != null ? JSON.parse(like) : initialValue
+            const item = window.localStorage.getItem(key)
+            return item !== null ? JSON.parse(item) : initialValue
         } catch (error) {
             return initialValue
         }
@@ -25,39 +26,22 @@ function useLocalStorage(key, initialValue) {
             console.log(error);
         }
     }
-    return [useLocalStorage, setLocalStorage]
+
+    //retornar los resultados 
+    return [values, setLocalStorage]
 }
 
 
 export const PhotoCard = ({ id, likes = 0, src = IMAGE_DEFAULT }) => {
     //referencia de los elememtos de los componentes
-    const element = useRef(null);
-    const [show, setShow] = useState(false);
-    const key = `like-${id}`;
-    const [liked, setLiked] = useLocalStorage(key, false);
+    //polify
+    const [show, element] = UsearScreen()
 
-    useEffect(function () {
-        //carga del poliffy
-        Promise.resolve(
-            typeof window.IntersectionObserver != 'undefined'
-                ? window.IntersectionObserver
-                : import('intersection-observer')
-        ).then(() => {
-            const observer = new
-                window.IntersectionObserver(function
-                    (entries) {
-                    const { isIntersecting } = entries[0]
-                    if (isIntersecting) {
-                        setShow(true)
-                    }
-                })
-            observer.observe(element.current)
-        })
-    }, [element])
+    const key = `like-${id}`;
+    const [likeds, setLiked] = useLocalStorage(key, false);
 
     //cambiat de icono
-    console.log(liked);
-    const Icon = liked ? MdFavorite : MdFavoriteBorder
+    const Icon = likeds ? MdFavorite : MdFavoriteBorder
 
     return (
         <Article ref={element}>
@@ -69,7 +53,7 @@ export const PhotoCard = ({ id, likes = 0, src = IMAGE_DEFAULT }) => {
                             <Img src={src} />
                         </ImgWrapper>
                     </a>
-                    <Button onClick={() => setLiked(!liked)}>
+                    <Button onClick={() => setLiked(!likeds)}>
                         <Icon size='30px' />
                         {likes} Gustos!!
                     </Button>
